% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{add_default_hooks}
\alias{add_default_hooks}
\title{default hooks for API}
\usage{
add_default_hooks(api)
}
\arguments{
\item{api}{a Plumber api object}
}
\value{
a Plumber api object with added hooks
}
\description{
Adds a default set of hooks (currently only
pre route) to the API. These hooks will be used for
logging interactions with the running api.
}
