% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{create_routes}
\alias{create_routes}
\title{Create named from files}
\usage{
create_routes(dir)
}
\arguments{
\item{dir}{directory containing end points}
}
\value{
a named vector of characters
}
\description{
For a given directory, create a named vector of files and endpoints
}
