% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{generate_api}
\alias{generate_api}
\title{Generate APIs}
\usage{
generate_api(routes, ...)
}
\arguments{
\item{routes}{A list of files. Element names are the plumber endpoints}

\item{...}{filters and/or envir args to pass to plumber::pr}
}
\value{
A Plumber object
}
\description{
routes is typically created by create_routes
}
