% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{get_internal_routes}
\alias{get_internal_routes}
\title{Get API Routes}
\usage{
get_internal_routes(path = ".")
}
\arguments{
\item{path}{Additional path added to extdata/api/routes/path}
}
\value{
character, a file path
}
\description{
Convenience function.
}
