% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{setup_logger}
\alias{setup_logger}
\title{Set up default logger}
\usage{
setup_logger(dir = "./API_logs")
}
\arguments{
\item{dir}{directory path for logs}
}
\description{
Creates a rotating file log using json format, see
\link[logger]{appender_file} for details.
}
